+++ example/src/NativeBannerExample.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { useState } from 'react';
 import { Platform, StyleSheet } from 'react-native';
 import { AdView, AdFormat } from '../../src/index';
 import type { AdInfo, AdLoadFailedInfo, AdRevenueInfo } from '../../src/index';
@@ -21,6 +22,10 @@ const NativeBannerExample = ({
     isProgrammaticBannerShowing,
     setIsNativeUIBannerShowing,
 }: Props) => {
+    const adaptiveBannerEnabled = true; // true by default
+
+    const [bannerSize, setBannerSize] = useState({});
+
     return (
         <>
             <AppButton
@@ -34,9 +39,13 @@ const NativeBannerExample = ({
                 <AdView
                     adUnitId={adUnitId}
                     adFormat={AdFormat.BANNER}
-                    style={styles.banner}
+                    style={{...styles.banner, ...bannerSize}}
+                    adaptiveBannerEnabled={adaptiveBannerEnabled}
                     onAdLoaded={(adInfo: AdInfo) => {
                         log('Banner ad loaded from ' + adInfo.networkName);
+                        if (adaptiveBannerEnabled) {
+                             setBannerSize({width: adInfo.size?.width, height: adInfo.size?.height});
+                        }
                     }}
                     onAdLoadFailed={(errorInfo: AdLoadFailedInfo) => {
                         log(
@@ -75,9 +84,10 @@ const styles = StyleSheet.create({
         // Set background color for banners to be fully functional
         backgroundColor: '#000000',
         position: 'absolute',
-        width: '100%',
+        alignSelf: 'center',
         // Automatically sized to 50 on phones and 90 on tablets. When adaptiveBannerEnabled is on,
         // sized to AppLovinMAX.getAdaptiveBannerHeightForWidth().
+        width: 'auto',
         height: 'auto',
         bottom: Platform.select({
             ios: 36, // For bottom safe area